boosted fault tolleranceto number format exceptions

+++ app/src/main/java/com/karmadeliworks/paymentprocessing/Utility.java
@@ -1,1 +1,1 @@
-package com.karmadeliworks.paymentprocessing;public class Utility {    static String flag = "";    static  Boolean alertSummoned = false;    static int amountSent = 0;}
+package com.karmadeliworks.paymentprocessing;public class Utility {    static String flag = "";    static  Boolean alertSummoned = false;    static float amountSent = 0f;    }
